Consolidate fmm tests
